refactor(site-spec): add support for sub-pages
Allow SiteSpecification.collections to map collections to pages
as well as sub-pages, and tweak the CLI tool, formsg-site-spec
and site-generator to take these into account.

When generating sub-pages, impose a hard limit of 26 sub-pages
(the number of letters in the English alphabet), so that we can
comfortably generate sub-pages without too much complexity.